style: :lipstick: corrected and added paging to improve user accessibility
By @AyelmerCorzoB

+++ src/main/java/com/ejercicio/startapp/Models/Customer.java
@@ -1,13 +1,13 @@
 package com.ejercicio.startapp.Models;
 
 public class Customer {
-    private Long id;
+    private int id;
     private String firstName;
     private String lastName;
     private String email;
     public Customer() {
     }
-    public Customer(Long id, String firstName, String lastName, String email) {
+    public Customer(int id, String firstName, String lastName, String email) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
@@ -18,10 +18,10 @@ public class Customer {
         this.lastName = lastName;
         this.email = email;
     }
-    public Long getId() {
+    public int getId() {
         return id;
     }
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
     public String getFirstName() {
